Change channel binary size to u8

+++ src/channels.rs
@@ -90,8 +90,8 @@ pub enum FrequencyStatus {
 
 #[derive(Debug, PartialEq, Clone, Eq)]
 pub enum WiFiChannel {
-    Channel2GHz(u16),
-    Channel5GHz(u16),
+    Channel2GHz(u8),
+    Channel5GHz(u8),
 }
 
 #[allow(clippy::incorrect_partial_ord_impl_on_ord_type)]
@@ -127,7 +127,7 @@ impl fmt::Display for WiFiChannel {
 }
 
 impl WiFiChannel {
-    pub fn new(channel: u16) -> Option<WiFiChannel> {
+    pub fn new(channel: u8) -> Option<WiFiChannel> {
         match channel {
             1..=14 => Some(WiFiChannel::Channel2GHz(channel)),
             34 | 36 | 38 | 40 | 42 | 44 | 46 | 48 | 50 | 52 | 54 | 56 | 58 | 60 | 62 | 64 | 100
@@ -138,7 +138,7 @@ impl WiFiChannel {
         }
     }
 
-    pub fn get_channel_number(&self) -> u16 {
+    pub fn get_channel_number(&self) -> u8 {
         match self {
             WiFiChannel::Channel2GHz(channel) => *channel,
             WiFiChannel::Channel5GHz(channel) => *channel,
